Feat/adicionando autenticacao bff

+++ ExpenseCore/client/Registration/src/application/ExcludeClient.ts
@@ -5,6 +5,7 @@ import { IUnitOfWorkApplication } from "./IUnitOfWorkApplication";
 import { ClientCreatedDto } from "./dto/ClientCreatedDto";
 import { InternalServerErrorException } from "../infra/exceptions/InternalServerErrorException";
 import { ExcludeClientDto } from "./dto/ExcludeClientDto";
+import { ClientRegistrationDto } from "./dto/ClientRegistrationDto";
 
 export class ExcludeClient implements IExcludeClient {
     constructor(
@@ -16,27 +17,33 @@ export class ExcludeClient implements IExcludeClient {
     ) { }
 
     public async execute(dto: ExcludeClientDto): Promise<void> {
+        let userId = ''
         try {
             this.logger.info(`ExcludeClient - Iniciando Step 5 para exclusao do cliente`)
             this.logger.info(`ExcludeClient - Recuperando Estado da execucao CreateClient`)
             const clientCreated = await this.stateManager.get<ClientCreatedDto>('ClientCreated')
-            if (clientCreated) {
+            const createClientDto = await this.stateManager.get<ClientRegistrationDto>('createClient')
+            if (clientCreated && createClientDto) {
+                userId = createClientDto.userId
                 await this.unitOfWork.startTransaction();
                 this.logger.info(`ExcludeClient - Recuperando cliente`)
                 const client = await this.repository.findClientById(clientCreated.clientId)
                 if (client) {
                     this.logger.info(`ExcludeClient - Excluindo cliente`)
                     await this.repository.delete(client)
+                    await this.unitOfWork.commit()
+                    this.logger.info('ExcludeClient - Publicando evento de erro na criacao do cliente')
+                    await this.queue.publish('client.events', 'client.registration.error', { 
+                        error: { 
+                            message: dto.error.message,
+                            statusCode: dto.error.status,
+                            timestamp: new Date().toISOString()
+                        },
+                        data: {
+                            userId
+                        }
+                    })
                 }
-                await this.unitOfWork.commit()
-                this.logger.info('ExcludeClient - Publicando evento de erro na criacao do cliente')
-                await this.queue.publish('client.events', 'client.registration.error', { 
-                    error: { 
-                        message: dto.error.message,
-                        status: dto.error.status
-                    } 
-                })
-                return
             } else {
                 this.logger.info('ExcludeClient - Estado não recuperado')
             }
@@ -50,7 +57,10 @@ export class ExcludeClient implements IExcludeClient {
                 error: { 
                     message: dto.error.message,
                     status: dto.error.status
-                } 
+                },
+                data: {
+                    userId
+                }
             })
         }
     }
